fixed cargo loop
